Make LoadingViewHolder a separate class
