+++ ProjectSource/app-library/src/main/java/com/kayako/sdk/android/k5/common/adapter/loadmorelist/EndlessRecyclerViewScrollAdapter.java
@@ -4,7 +4,6 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ProgressBar;
 
 import com.kayako.sdk.android.k5.R;
 import com.kayako.sdk.android.k5.common.adapter.BaseListItem;
@@ -109,14 +108,7 @@ public abstract class EndlessRecyclerViewScrollAdapter extends RecyclerView.Adap
         return getData().size();
     }
 
-    public class LoadingViewHolder extends RecyclerView.ViewHolder {
-        public ProgressBar mProgressBar;
 
-        public LoadingViewHolder(View view) {
-            super(view);
-            mProgressBar = (ProgressBar) view.findViewById(R.id.ko__view_loading);
-        }
-    }
 
     public interface OnLoadMoreListener {
         void loadMoreItems();
